refactor: move groupware stores to dedicated groupware folder
