+++ packages/web-app-mail/src/components/MailDetails.vue
@@ -57,7 +57,7 @@ import { useGettext } from 'vue3-gettext'
 import { buildMailBody } from '../helpers'
 import MailAttachmentList from './MailAttachmentList.vue'
 import MailIndicators from './MailIndicators.vue'
-import { AppLoadingSpinner, useAccountsStore } from '@opencloud-eu/web-pkg'
+import { AppLoadingSpinner, useGroupwareAccountsStore } from '@opencloud-eu/web-pkg'
 import MailAppointmentList from './MailAppointmentList.vue'
 import { useLoadMail } from '../composables/useLoadMail'
 import { useMailsStore } from '../composables/piniaStores/mails'
@@ -65,7 +65,7 @@ import { storeToRefs } from 'pinia'
 
 const { isLoading } = useLoadMail()
 const { current: currentLanguage } = useGettext()
-const accountsStore = useAccountsStore()
+const accountsStore = useGroupwareAccountsStore()
 const mailsStore = useMailsStore()
 const { currentAccount } = storeToRefs(accountsStore)
 const { currentMail } = storeToRefs(mailsStore)
