+++ packages/web-pkg/src/composables/piniaStores/groupware/index.ts
@@ -1,1 +1,3 @@
 export * from './accounts'
+export * from './config'
+export * from './types'
