+++ packages/web-pkg/src/composables/piniaStores/groupware/types.ts
