--- packages/web-pkg/src/composables/piniaStores/groupwareConfig/index.ts
@@ -1,2 +1,0 @@
-export * from './groupwareConfig'
-export * from './types'
