+++ packages/web-pkg/src/composables/piniaStores/index.ts
@@ -6,7 +6,6 @@ export * from './clipboard'
 export * from './config'
 export * from './extensionRegistry'
 export * from './groupware'
-export * from './groupwareConfig'
 export * from './messages'
 export * from './modals'
 export * from './resources'
